Breakout the Exception classes in Properties.cs to their own file
